Fix message format

+++ cabot_can/src/can_all_node.cpp
@@ -86,10 +86,10 @@ public:
     pressure_pub_ = this->create_publisher<sensor_msgs::msg::FluidPressure>("pressure", 2);
     calibration_pub_ = this->create_publisher<std_msgs::msg::Int32MultiArray>("calibration", 50);
     tact_pub_ = this->create_publisher<std_msgs::msg::Int8>("pushed", 50);
-    capacitive_touch_pub_ = this->create_publisher<std_msgs::msg::Int8>("capacitive/touch", 50);
-    tof_touch_pub_ = this->create_publisher<std_msgs::msg::Int8>("tof/touch", 50);
-    touch_pub_ = this->create_publisher<std_msgs::msg::Int8>("touch", 50);
-    capacitive_touch_raw_pub_ = this->create_publisher<std_msgs::msg::Int8>("capacitive/touch_raw", 50);
+    capacitive_touch_pub_ = this->create_publisher<std_msgs::msg::Int16>("capacitive/touch", 50);
+    tof_touch_pub_ = this->create_publisher<std_msgs::msg::Int16>("tof/touch", 50);
+    touch_pub_ = this->create_publisher<std_msgs::msg::Int16>("touch", 50);
+    capacitive_touch_raw_pub_ = this->create_publisher<std_msgs::msg::Int16>("capacitive/touch_raw", 50);
     tof_touch_raw_pub_ = this->create_publisher<std_msgs::msg::Int16>("tof/touch_raw", 50);
     servo_pos_pub_ = this->create_publisher<std_msgs::msg::Int16>("servo_pos", 50);
     vibrator_1_sub_ = this->create_subscription<std_msgs::msg::UInt8>("vibrator1", 10,[this](const std_msgs::msg::UInt8::SharedPtr msg) {this->subscribeVibratorData(msg, 1);});
@@ -383,27 +383,25 @@ private:
   void publishTouchData(const struct can_frame &frame) {
     if (frame.can_id == CanId::TOUCH_CAN_ID && frame.can_dlc >= 4) {
       int8_t capacitive_touch = frame.data[3];
-      std_msgs::msg::Int8 touch_msg;
+      std_msgs::msg::Int16 touch_msg;
       touch_msg.data = capacitive_touch;
       capacitive_touch_pub_->publish(touch_msg);
       touch_pub_->publish(touch_msg);
-      
       int8_t capacitive_touch_raw = frame.data[2];
-      std_msgs::msg::Int8 capacitive_touch_raw_msg;
+      std_msgs::msg::Int16 capacitive_touch_raw_msg;
       capacitive_touch_raw_msg.data = capacitive_touch_raw;
       capacitive_touch_raw_pub_->publish(capacitive_touch_raw_msg);
-      
       int16_t tof_touch_raw = (((uint16_t)frame.data[1]) << 8) | ((uint16_t)frame.data[0]);
       std_msgs::msg::Int16 tof_raw_msg;
       tof_raw_msg.data = tof_touch_raw;
       tof_touch_raw_pub_->publish(tof_raw_msg);
-      int8_t tof_touch = 0;
+      int16_t tof_touch = 0;
       if (tof_touch_raw >= 16 && tof_touch_raw <= 25) {
           tof_touch = 1;
       } else {
           tof_touch = 0;
       }
-      std_msgs::msg::Int8 tof_touch_msg;
+      std_msgs::msg::Int16 tof_touch_msg;
       tof_touch_msg.data = tof_touch;
       tof_touch_pub_->publish(tof_touch_msg);  
     }
@@ -537,11 +535,11 @@ private:
   rclcpp::Publisher<sensor_msgs::msg::FluidPressure>::SharedPtr pressure_pub_;
   rclcpp::Publisher< std_msgs::msg::Int32MultiArray>::SharedPtr calibration_pub_;
   rclcpp::Publisher<std_msgs::msg::Int8>::SharedPtr tact_pub_;
-  rclcpp::Publisher<std_msgs::msg::Int8>::SharedPtr capacitive_touch_pub_;
-  rclcpp::Publisher<std_msgs::msg::Int8>::SharedPtr capacitive_touch_raw_pub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr capacitive_touch_pub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr capacitive_touch_raw_pub_;
   rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr tof_touch_raw_pub_;
-  rclcpp::Publisher<std_msgs::msg::Int8>::SharedPtr tof_touch_pub_;
-  rclcpp::Publisher<std_msgs::msg::Int8>::SharedPtr touch_pub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr tof_touch_pub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr touch_pub_;
   rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr servo_pos_pub_;
   rclcpp::Subscription<std_msgs::msg::UInt8>::SharedPtr vibrator_1_sub_;
   rclcpp::Subscription<std_msgs::msg::UInt8>::SharedPtr vibrator_3_sub_;
